add tabs component
