+++ src/legacy/Tab/index.ts
@@ -1,0 +1,3 @@
+export { TabsBar as LegacyTabsBar } from './TabsBar';
+export { Tab as LegacyTab } from './Tab';
+export { TabContent as LegacyTabContent } from './TabContent';
