+++ src/legacy/index.ts
@@ -2,3 +2,4 @@ export * from './Button';
 export * from './Checkbox';
 export * from './Switch';
 export { LegacyTable as LegacyTable } from './Table';
+export * from './Tab';
